hslua-oo: move code into subdirectory
