+++ hslua-objectorientation/hslua-objectorientation.cabal
@@ -58,6 +58,8 @@ common common-options
 library
   import:              common-options
   exposed-modules:     HsLua.ObjectOrientation
+                     , HsLua.ObjectOrientation.Generic
+                     , HsLua.ObjectOrientation.ListType
                      , HsLua.ObjectOrientation.Operation
   hs-source-dirs:      src
   build-depends:       containers           >= 0.5.9  && < 0.9
